refactor(kitchenpos) 패키지 구성 피드백 반영
