+++ src/main/java/kitchenpos/menu/domain/JpaMenuGroupRepository.java
@@ -1,6 +1,5 @@
-package kitchenpos.menu.repository;
+package kitchenpos.menu.domain;
 
-import kitchenpos.menu.domain.MenuGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
