+++ src/main/java/kitchenpos/menu/domain/JpaMenuRepository.java
@@ -1,6 +1,5 @@
-package kitchenpos.menu.repository;
+package kitchenpos.menu.domain;
 
-import kitchenpos.menu.domain.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
