+++ src/main/java/kitchenpos/order/domain/JpaOrderRepository.java
@@ -1,6 +1,5 @@
-package kitchenpos.order.repository;
+package kitchenpos.order.domain;
 
-import kitchenpos.order.domain.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
