+++ src/main/java/kitchenpos/ordertable/domain/JpaOrderTableRepository.java
@@ -1,6 +1,5 @@
-package kitchenpos.ordertable.repository;
+package kitchenpos.ordertable.domain;
 
-import kitchenpos.ordertable.domain.OrderTable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
