+++ src/main/java/kitchenpos/ordertable/domain/OrderTableRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.ordertable.repository;
+package kitchenpos.ordertable.domain;
 
 import kitchenpos.ordertable.domain.OrderTable;
